Shaked: Refactored arm subsystem
