--- src/main/kotlin/frc/robot/commands/arm/ArmCommandsConstants.kt
@@ -1,3 +1,0 @@
-package frc.robot.commands.arm
-
-const val DEFAULT_OUTPUT = 0.0
