+++ src/main/kotlin/frc/robot/subsystems/arm/ArmConstants.kt
@@ -1,5 +1,7 @@
 package frc.robot.subsystems.arm
 
 object ArmConstants {
-    const val PERCENT_OUTPUT = 0.0
+    const val DEFAULT_OUTPUT = 0.0
+    const val STATOR_CURRENT_AT_REVERSE_LIMIT = 0.0
+    const val AT_REVERSE_LIMIT_TIME = 0.0
 }
