+++ src/main/kotlin/frc/robot/subsystems/arm/ArmSubsystem.kt
@@ -3,29 +3,48 @@ package frc.robot.subsystems.arm
 import com.ctre.phoenix.motorcontrol.NeutralMode
 import com.hamosad1657.lib.motors.HaTalonSRX
 import edu.wpi.first.wpilibj.DigitalInput
+import edu.wpi.first.wpilibj.Timer
 import edu.wpi.first.wpilibj2.command.SubsystemBase
 import frc.robot.RobotMap.Arm as ArmMap
 
 object ArmSubsystem : SubsystemBase() {
-    // TODO: Check if limits are wired normally false pr normally false.
+    // TODO: Check if limit is wired normally false or normally false.
 
-    private val motor = HaTalonSRX(ArmMap.MOTOR_ID)
+    private val leftMotor = HaTalonSRX(ArmMap.MOTOR_ID)
+    private val rightMotor = HaTalonSRX(ArmMap.MOTOR_ID)
 
     init {
-        motor.inverted = false // TODO: Verify positive output opens arm.
+        leftMotor.inverted = false // TODO: Verify positive output opens arm.
+        rightMotor.inverted = false
+        rightMotor.follow(leftMotor)
     }
 
     private val forwardLimit = DigitalInput(ArmMap.FORWARD_LIMIT_CHANNEL)
-    private val reverseLimit = DigitalInput(ArmMap.REVERSE_LIMIT_CHANNEL)
+
+    private val reverseLimitTimer = Timer().apply {
+        start()
+    }
 
     fun setNeutralMode(neutralMode: NeutralMode) {
-        motor.setNeutralMode(neutralMode)
+        leftMotor.setNeutralMode(neutralMode)
+        rightMotor.setNeutralMode(neutralMode)
     }
 
     fun set(percentOutput: Double) {
-        motor.set(percentOutput)
+        leftMotor.set(percentOutput)
     }
 
-    fun isAtForwardLimit() = forwardLimit.get()
-    fun isAtReverseLimit() = reverseLimit.get()
+    val isAtForwardLimit get() = forwardLimit.get()
+
+    val isAtReverseLimit: Boolean
+        get() {
+            if ((leftMotor.statorCurrent >= ArmConstants.STATOR_CURRENT_AT_REVERSE_LIMIT) &&
+                reverseLimitTimer.hasElapsed(ArmConstants.AT_REVERSE_LIMIT_TIME)
+            ) {
+                return true
+            } else {
+                reverseLimitTimer.restart()
+                return false
+            }
+        }
 }
